Added double casts to warrior DPS calculation, cleaned up code.

+++ Characters/Warrior.cs
@@ -99,7 +99,7 @@ namespace RPG_Characters
         {
             if (Equipment[Slot.SLOT_WEAPON] != null)
             {
-                return Equipment[Slot.SLOT_WEAPON].GetDPS() + 1 + (BasePrimaryAttributes.Strength + TotalPrimaryAttributes.Strength) / 100;
+                return (double)((double)Equipment[Slot.SLOT_WEAPON].GetDPS() + 1.0 + ((double)BasePrimaryAttributes.Strength + (double)TotalPrimaryAttributes.Strength) / 100);
             }
             return 1 + ((double)(BasePrimaryAttributes.Strength + (double)TotalPrimaryAttributes.Strength) / 100);
         }
